Switched from singularity to apptainer RPM
dwd recommended to use the default apptainer installation and configuration. This allows to run inside another singularity/apptainer but requires unprivileged user namespaces enabled

If unprivileged user namespaces are disabled also  setuid-root singularity/apptainer needs to be installed, but this would break running inside another singularity/apptainer. So it is not done. To be considered if needed in the future